refactor: unify site health check implementations
