+++ src/bot/commands/check.ts
@@ -1,6 +1,7 @@
 import axiosCheck from "../../methods/axios";
 import browserCheck from "../../methods/browser";
 import { WebModel } from "../../modules/sqlModel";
+import { checkAll } from "../../utils/checkAll";
 import { MessageProcessor } from "../adapters/botAdapter";
 
 export const check: MessageProcessor = async (ctx) => {
@@ -16,10 +17,9 @@ export const check: MessageProcessor = async (ctx) => {
 			browserCheck();
 		} else if (input === "all") {
 			ctx.reply("巡查任务已启动，请稍后查看报告捏 ~");
-			await axiosCheck();
-			browserCheck();
+			checkAll();
 		} else {
-			ctx.reply("方式无效，当前可选方式：axios, browser 喵");
+			ctx.reply("方式无效，当前可选方式：axios, browser, all 喵");
 			return;
 		}
 	} else {
